Update iptables-lockdown.sh
A bash script used to lockdown the IPTables local firewall on Linux to ensure it is only connected to from trusted machine. This is often used in situations where your vulnerable lab server has to be on the same network as other assets or networks you do not trust or want to expose a vulnerable machine to (e.g. Bridge mode in VM networking)

The script does the following:
- Backup existing firewall rules
- Flush or clear all firewall existing rules
- Temporarily change default policy to ACCEPT to prevent locking oneself out
- Add whitelisted IPs to INPUT/OUTPUT
- Log to a log file
- Backup to rules.v4